+++ imports/ui/proposals/Proposal.jsx
@@ -14,7 +14,7 @@ import i18n from 'meteor/universe:i18n';
 import { Meteor } from 'meteor/meteor';
 import Coin from '/both/utils/coins.js';
 import TimeStamp from '../components/TimeStamp.jsx';
-// import { ProposalActionButtons } from '../ledger/LedgerActions.jsx';
+import { ProposalActionButtons } from '../ledger/LedgerActions.jsx';
 import voca from 'voca';
 
 const T = i18n.createComponent();
@@ -293,7 +293,7 @@ export default class Proposal extends Component{
                         <Row className="mb-2 border-top">
                             <Col md={3} className="label"><T>proposals.proposalID</T></Col>
                             <Col md={this.state.user?6:9} className="value">{this.props.proposal.proposalId}</Col>
-                            {/* {this.state.user?<Col md={3}><ProposalActionButtons history={this.props.history} proposalId={proposalId}/></Col>:null} */}
+                            {this.state.user?<Col md={3}><ProposalActionButtons history={this.props.history} proposalId={proposalId}/></Col>:null}
                         </Row>
                         <Row className="mb-2 border-top">
                             <Col md={3} className="label"><T>proposals.proposer</T></Col>